folder icons changed in save to modal
